docs: improve template docs
* docs: remove old docs

* docs: restructure main readme

* docs: extend let docs

* Update libs/template/let/src/lib/docs/Readme.md

* Update libs/template/let/src/lib/docs/Readme.md

* Update libs/template/let/src/lib/docs/Readme.md

* docs: start testing section

* docs: add template trigger to rxLet

* docs(template): improve rx-let docs

* docs: add concepts

* docs: add concepts

* docs(template): improve rx-let docs

* docs: arrange concepts

* docs: add concepts

* docs(template): update let docs

* Update libs/template/let/src/lib/docs/Readme.md

* docs(template): update let docs

* docs(template): update let docs

* docs: introduce new folder for perf issues

* docs(template): update let docs

* docs: add docs

* docs(template): update let docs

* docs(template): update template docs

* docs(template): fix typo

* docs(template): fix typos

* docs: delete perf folder

* docs: restructure files

* docs: add resources and donwsides

* docs: add tips

* docs(template): improve rxFor docs

* Update libs/template/for/src/lib/docs/Readme.md

* refactor(docs): move template docs to docs app

* refactor(docs): bring back concepts.md

* refactor(demos): fix import path

* refactor(docs): adjust navigation structure

* refactor(docs): improve structure of let and for docs

* docs(template): align jsdocs with docs app

* docs(template): align jsdocs with docs app
