+++ apps/demos/src/app/features/tutorials/basics/1-setup/setup.start.component.ts
@@ -1,9 +1,19 @@
-import { ChangeDetectionStrategy, Component, Input, OnDestroy, OnInit, Output } from '@angular/core';
-import { interval, Subject, Subscription } from 'rxjs';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  Input,
+  OnDestroy,
+  OnInit,
+  Output,
+} from '@angular/core';
+import { filter, interval, merge, Observable, Subject, Subscription, switchMap, takeUntil } from 'rxjs';
 import { map, startWith, tap } from 'rxjs/operators';
 import { ListServerItem, ListService } from '../data-access/list-resource';
 //👇 1- import RxState
 import { RxState } from '@rx-angular/state';
+import { RxActionFactory } from '@rx-angular/state/actions';
+import { RxEffects } from '@rx-angular/state/effects';
+
 //👇 2- define a component state
 interface ComponentState {
   refreshInterval: number;
@@ -20,28 +30,28 @@ export interface DemoBasicsItem {
 const initComponentState = {
   refreshInterval: 10000,
   listExpanded: false,
-  list: []
+  list: [],
 };
 
 @Component({
   selector: 'rxa-setup-start',
   //👇 Render the model property of the component
   template: `
-    <h3>
-      Setup
-    </h3>
+    <h3>Setup</h3>
     <mat-expansion-panel
-      (expandedChange)="listExpanded = $event; listExpandedChanges.next($event)"
-      [expanded]="listExpanded">
+      *rxLet="vm$; let vm"
+      (expandedChange)="
+        vm.listExpanded = $event; ui.listExpandedChanges($event)
+      "
+      [expanded]="listExpanded"
+    >
       <mat-expansion-panel-header class="list">
         <mat-progress-bar *ngIf="false" [mode]="'query'"></mat-progress-bar>
-        <mat-panel-title>
-          List
-        </mat-panel-title>
+        <mat-panel-title> List</mat-panel-title>
         <mat-panel-description>
           <span
-          >{{ (storeList$ | async)?.length }} Repositories Updated every:
-            {{ _refreshInterval }} ms
+            >{{ vm.list.length }} Repositories Updated every:
+            {{ vm.refreshInterval }} ms
           </span>
         </mat-panel-description>
       </mat-expansion-panel-header>
@@ -49,20 +59,18 @@ const initComponentState = {
       <button
         mat-raised-button
         color="primary"
-        (click)="onRefreshClicks($event)"
+        (click)="ui.refreshClicks($event)"
       >
         Refresh List
       </button>
 
-      <ng-container *ngIf="storeList$ | async as list">
-        <div *ngIf="list?.length; else noList">
-          <mat-list>
-            <mat-list-item *ngFor="let item of list">
-              {{ item.name }}
-            </mat-list-item>
-          </mat-list>
-        </div>
-      </ng-container>
+      <div *ngIf="vm.list.length; else noList">
+        <mat-list>
+          <mat-list-item *ngFor="let item of list">
+            {{ item.name }}
+          </mat-list-item>
+        </mat-list>
+      </div>
 
       <ng-template #noList>
         <mat-card>No list given!</mat-card>
@@ -84,58 +92,53 @@ const initComponentState = {
       .list .mat-expansion-panel-content .mat-expansion-panel-body {
         padding-top: 10px;
       }
-    `
+    `,
   ],
-  changeDetection: ChangeDetectionStrategy.OnPush
+  providers: [RxActionFactory, RxState, RxEffects],
+  changeDetection: ChangeDetectionStrategy.OnPush,
 })
-//👇 3- extend the component
-export class SetupStart implements OnInit, OnDestroy {
-  //👇 Set up the model property of the component
-  intervalSubscription = new Subscription();
-  listExpandedChanges = new Subject<boolean>();
-  storeList$ = this.listService.list$.pipe(
-    map(this.parseListItems),
-    startWith(initComponentState.list)
-  );
+export class SetupStart {
+  ui = this.rxActions.create();
+  vm$ = this.model.select();
 
-  _refreshInterval: number = initComponentState.refreshInterval;
   @Input()
-  set refreshInterval(refreshInterval: number) {
-    if (refreshInterval > 4000) {
-      this._refreshInterval = refreshInterval;
-      this.resetRefreshTick();
-    }
+  set refreshInterval(refreshInterval$: Observable<number>) {
+    this.model.connect(
+      'refreshInterval',
+      refreshInterval$.pipe(filter(rI => rI > 4000))
+    );
   }
 
-  listExpanded: boolean = initComponentState.listExpanded;
   @Output()
-  listExpandedChange = this.listExpandedChanges;
+  listExpandedChange = this.ui.listExpandedChanges$;
+
+  autoTrigger$ = this.model.select('refreshInterval')
+    .pipe(switchMap(ms => interval(ms)));
+
+  fetchEffect = (_) => this.listService.refetchList();
+
+
   constructor(
-    private listService: ListService
+    private listService: ListService,
+    private model: RxState<ComponentState>,
+    private ef: RxEffects,
+    private rxActions: RxActionFactory<{
+      listExpandedChanges: boolean,
+      refreshClicks: undefined
+    }>
   ) {
-    //👇 Always call super() first in the constructor
+    this.model.set(initComponentState);
 
-    //👇 Call set() to initialize the state
+    this.model.connect('list', this.listService.list$
+      .pipe(map(this.parseListItems))
+    );
 
-  }
 
-  ngOnDestroy(): void {
-    this.intervalSubscription.unsubscribe();
-  }
+    this.ef.register(
+      merge(this.autoTrigger$, this.ui.listExpandedChanges$),
+      this.fetchEffect
+    );
 
-  ngOnInit(): void {
-    this.resetRefreshTick();
-  }
-
-  resetRefreshTick() {
-    this.intervalSubscription.unsubscribe();
-    this.intervalSubscription = interval(this._refreshInterval)
-      .pipe(tap((_) => this.listService.refetchList()))
-      .subscribe();
-  }
-
-  onRefreshClicks(event) {
-    this.listService.refetchList();
   }
 
   parseListItems(l: ListServerItem[]): DemoBasicsItem[] {
