+++ apps/docs/docs/cdk/render-strategies/_docs.md
@@ -15,7 +15,7 @@ Furthermore, they provide new ways of explicitly tying truly push-based state ma
 
 A strategy exposes the work to perform (e.g. `cdRef#markForCheck`, `cdRef#detectChanges`) as well as the scheduling mechanism to developers for configuration & customization via the interface `RxStrategyCredentials`.
 
-`Directive`s, `Service`s or `Component`s of your application can make use of these strategies as an easy API for the key [concepts](../../template/concepts/concepts.md) of rendering performance.
+`Directive`s, `Service`s or `Component`s of your application can make use of these strategies as an easy API for the key [concepts](../../template/concepts/coalescing-scheduling.md) of rendering performance.
 
 This architecture enables modern features like:
 
@@ -70,18 +70,18 @@ The sub-package provides the following features:
 
 **[Available Strategies:](render-strategies/strategies)**
 
-| Name             | Priority | Render Method     | Scheduling              | Render Deadline |
-| ---------------- | -------- | ----------------- | ----------------------- | --------------- |
-| `"noop"`         | ❌       | - `noop`          | ❌                      | N/A             |
-| `"native"`       | ❌       | ⮁ `markForCheck`  | `requestAnimationFrame` | N/A             |
-| `"global"`       | ❌       | ⮁ `ɵmarkDirty`    | `requestAnimationFrame` | N/A             |
-| `"local"`        | ❌       | 🠗 `detectChanges` | `requestAnimationFrame` | N/A             |
-|                  |          |                   |                         |                 |
+| Name             | Priority | Render Method      | Scheduling              | Render Deadline |
+|------------------|----------|--------------------|-------------------------|-----------------|
+| `"noop"`         | ❌        | - `noop`           | ❌                       | N/A             |
+| `"native"`       | ❌        | ⮁ `markForCheck`   | `requestAnimationFrame` | N/A             |
+| `"global"`       | ❌        | ⮁ `ɵmarkDirty`     | `requestAnimationFrame` | N/A             |
+| `"local"`        | ❌        | 🠗 `detectChanges` | `requestAnimationFrame` | N/A             |
+|                  |          |                    |                         |                 |
 | `"immediate"`    | 1        | 🠗 `detectChanges` | `postMessage`           | 0ms             |
 | `"userBlocking"` | 2        | 🠗 `detectChanges` | `postMessage`           | 250ms           |
 | `"normal"`       | 3        | 🠗 `detectChanges` | `postMessage`           | 5000ms          |
 | `"low"`          | 4        | 🠗 `detectChanges` | `postMessage`           | 10000ms         |
-| `"idle"`         | 5        | 🠗 `detectChanges` | `postMessage`           | ❌              |
+| `"idle"`         | 5        | 🠗 `detectChanges` | `postMessage`           | ❌               |
 
 **Zone notification configuration with patchZone property**
