+++ apps/docs/docs/template/api/unpatch-directive.md
@@ -1,5 +1,7 @@
 ---
-sidebar_position: 1
+sidebar_label: 'Unpatch'
+sidebar_position: 3
+title: 'Unpatch'
 ---
 
 # UnpatchDirective
@@ -16,7 +18,7 @@ The current way of binding events to DOM:
 ```
 
 The problem is that every event registered via `()`, e.g. `(mousemove)` (or custom `@Output()`)
-marks the component and all its ancestors as dirty and re-renders the whole component tree. [read more about this here](../concepts/performance-issues.md)
+marks the component and all its ancestors as dirty and re-renders the whole component tree. [read more about this here](../performance-issues/rendering-issues-in-angular.md)
 
 So even if your eventListener is not related to any change at all, your app will re-render the whole component tree.
 This can lead to very bad user experiences, especially if you work with frequently fired events such as `mousemove`.
