+++ apps/docs/docs/template/concepts/_category_.json
@@ -1,4 +1,9 @@
 {
   "label": "Concepts",
-  "position": 1
+  "position": 110,
+  "link": {
+    "type": "generated-index",
+    "title": "Concepts",
+    "slug": "/template/concepts"
+  }
 }
