+++ apps/docs/docs/template/performance-issues/_category_.json
@@ -1,0 +1,9 @@
+{
+  "label": "Performance Issues",
+  "position": 120,
+  "link": {
+    "type": "generated-index",
+    "title": "Performance Issues",
+    "slug": "/template/performance-issues"
+  }
+}
