+++ libs/template/experimental/viewport-prio/viewport-prio.md
@@ -1,0 +1,5 @@
+# 🧪 ViewportPrioDirective
+
+This directive limits renderings to only visible components. Should be used together with Noop strategy.
+
+_more info coming soon_
