+++ libs/template/unpatch/src/lib/README.md
@@ -1,0 +1,29 @@
+# @rx-angular/template/unpatch
+
+[![npm](https://img.shields.io/npm/v/%40rx-angular%2Ftemplate.svg)](https://www.npmjs.com/package/%40rx-angular%2Ftemplate)
+![rx-angular CI](https://github.com/rx-angular/rx-angular/workflows/rx-angular%20CI/badge.svg?branch=master)
+[![Coverage Status](https://raw.githubusercontent.com/rx-angular/rx-angular/github-pages/docs/test-coverage/template/jest-coverage-badge.svg)](https://rx-angular.github.io/rx-angular/test-coverage/template/lcov-report/index.html)
+
+## This package provides a quick way to create an action channel and some configuration options.
+
+
+## Key features
+
+- ✅ Fully Typed
+
+## Demos:
+
+- [⚡ GitHub](https://github.com/rx-angular/rx-angular/blob/main/apps/demos/src/app/features/template/unpatch)
+
+## Install
+
+```bash
+npm install --save @rx-angular/template
+# or
+yarn add @rx-angular/template
+```
+
+## Documentation
+
+- [RxLet](https://rx-angular.io/docs/template/api/unpatch-directive)
+
